+++ src/features/client/mocks/categoriesMock.js
