+++ src/features/client/mocks/restaurantsMock.js
@@ -1,5 +1,6 @@
 export const restaurants = [
   {
+    id: 1,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -11,6 +12,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 2,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -22,6 +24,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 3,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -33,6 +36,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 4,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -44,6 +48,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 5,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -55,6 +60,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 6,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -66,6 +72,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 7,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -77,6 +84,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 8,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -88,6 +96,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 9,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -99,6 +108,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 10,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -110,6 +120,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 11,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
@@ -121,6 +132,7 @@ export const restaurants = [
       'https://images.unsplash.com/photo-1498837167922-ddd27525d352?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
   },
   {
+    id: 12,
     restaurantName: 'Jammi',
     restaurantAddress: 'Tauro Kalnas',
     restaurantBio: 'Fast food restaurant',
